Move StableId wrapper to common db package
In preparation for adding stable IDs to entities that aren't related to the
document producer, move the `StableId` wrapper to `com.terraformation.backend.db`.
