+++ src/main/kotlin/com/terraformation/backend/accelerator/variables/VariableValuesHelper.kt
@@ -1,5 +1,6 @@
 package com.terraformation.backend.accelerator.variables
 
+import com.terraformation.backend.db.StableId
 import com.terraformation.backend.db.default_schema.ProjectId
 import com.terraformation.backend.db.docprod.VariableId
 import com.terraformation.backend.db.docprod.VariableValueId
@@ -19,7 +20,6 @@ import com.terraformation.backend.documentproducer.model.NewSelectValue
 import com.terraformation.backend.documentproducer.model.NewTextValue
 import com.terraformation.backend.documentproducer.model.NumberVariable
 import com.terraformation.backend.documentproducer.model.SelectVariable
-import com.terraformation.backend.documentproducer.model.StableId
 import com.terraformation.backend.documentproducer.model.TextVariable
 import com.terraformation.backend.documentproducer.model.UpdateValueOperation
 import com.terraformation.backend.documentproducer.model.ValueOperation
