+++ src/main/kotlin/com/terraformation/backend/documentproducer/model/StableIds.kt
@@ -1,5 +1,6 @@
 package com.terraformation.backend.documentproducer.model
 
+import com.terraformation.backend.db.StableId
 import com.terraformation.backend.db.default_schema.LandUseModelType
 
 object StableIds {
